requirements.txt
# Raspberry Pi Camera Requirements
picamera2>=0.3.12
opencv-python>=4.5.0
numpy>=1.21.0


# Requirements for GC9A01 display test
# Note: Adafruit CircuitPython libraries are not available via pip
# Install them using: sudo apt install python3-pil python3-rpi.gpio
Pillow>=9.0.0
RPi.GPIO>=0.7.0
spidev>=3.5